+++ back-end/database/migrations/2024_04_02_195022_create_fournisseurs_table.php
@@ -8,7 +8,7 @@ return new class extends Migration {
     public function up()
     {
         Schema::create('fournisseur', function (Blueprint $table) {
-            $table->unsignedBigInteger('idFour');
+            $table->unsignedInteger('idFour');
 
             $table->primary('idFour');  // Use id() for auto-incrementing primary key
             $table->string('nom', 255)->collate('utf8mb4_unicode_ci');
