+++ back-end/database/migrations/2024_04_02_195026_create_messageries_table.php
@@ -9,7 +9,7 @@ return new class extends Migration
     public function up()
     {
         Schema::create('messagerie', function (Blueprint $table) {
-            $table->unsignedBigInteger('idmsg');  // Use id() for auto-incrementing primary key
+            $table->unsignedInteger('idmsg');  // Use id() for auto-incrementing primary key
             $table->primary('idmsg');  // Use id() for auto-incrementing primary key
             $table->string('mobile', 255)->collate('utf8mb4_unicode_ci');
             $table->text('wsms')->collate('utf8mb4_unicode_ci');
